refactor(client): centralize typed submission mutation documents
Colocating the mutation documents in their components bought little on
its own: consumers import the generated *Document either way, and the
real fragment/component restructuring is a separate effort. Move the
nine typed graphql() operations into a typed section at the top of
mutations.ts so the components stay light, and keep migrating legacy
gql exports into that section as their consumers adopt typed documents.

The userListItem fragment stays colocated on UserListItem — the roster
mutations select it, so the component keeps owning what it renders.

codegen, eslint, vue-tsc, and the full vitest suite (506) pass.
