+++ client/src/components/AssignedSubmissionUsers.vue
@@ -66,72 +66,6 @@
   </section>
 </template>
 
-<script lang="ts">
-import { graphql } from "src/graphql/generated"
-
-// These mutations return users that render in a UserList, so they include
-// the UserListItem-owned `userListItem` fragment.
-graphql(`
-  mutation UpdateSubmissionReviewers(
-    $id: ID!
-    $connect: [ID!]
-    $disconnect: [ID!]
-  ) {
-    updateSubmissionReviewers(
-      input: {
-        id: $id
-        reviewers: { connect: $connect, disconnect: $disconnect }
-      }
-    ) {
-      id
-      reviewers {
-        ...userListItem
-      }
-    }
-  }
-`)
-
-graphql(`
-  mutation UpdateSubmissionReviewCoordinators(
-    $id: ID!
-    $connect: [ID!]
-    $disconnect: [ID!]
-  ) {
-    updateSubmissionReviewCoordinators(
-      input: {
-        id: $id
-        review_coordinators: { connect: $connect, disconnect: $disconnect }
-      }
-    ) {
-      id
-      review_coordinators {
-        ...userListItem
-      }
-    }
-  }
-`)
-
-graphql(`
-  mutation UpdateSubmissionSubmitters(
-    $id: ID!
-    $connect: [ID!]
-    $disconnect: [ID!]
-  ) {
-    updateSubmissionSubmitters(
-      input: {
-        id: $id
-        submitters: { connect: $connect, disconnect: $disconnect }
-      }
-    ) {
-      id
-      submitters {
-        ...userListItem
-      }
-    }
-  }
-`)
-</script>
-
 <script setup lang="ts">
 import ReinviteUserDialog from "./dialogs/ReinviteUserDialog.vue"
 import FindUserSelect from "./forms/FindUserSelect.vue"
