+++ client/src/components/forms/CommentEditor.vue
@@ -78,143 +78,6 @@
   </q-form>
 </template>
 
-<script lang="ts">
-import { graphql } from "src/graphql/generated"
-
-// commentFields / relatedUserFields resolve from the shared fragment
-// registry (src/graphql/fragments.ts). Colocating the comment fragments
-// onto their rendering components is a separate follow-up.
-graphql(`
-  mutation CreateOverallComment(
-    $submission_id: ID!
-    $content: String!
-    $reply_to_id: ID
-    $parent_id: ID
-  ) {
-    createOverallComment(
-      input: {
-        submission_id: $submission_id
-        content: $content
-        reply_to_id: $reply_to_id
-        parent_id: $parent_id
-      }
-    ) {
-      id
-      overall_comments(trashed: WITH) {
-        ...commentFields
-        replies(trashed: WITH) {
-          ...commentFields
-          parent_id
-          reply_to_id
-        }
-      }
-    }
-  }
-`)
-
-graphql(`
-  mutation CreateInlineComment(
-    $submission_id: ID!
-    $content: String!
-    $from: Int
-    $to: Int
-    $style_criteria: [ID!]
-    $reply_to_id: ID
-    $parent_id: ID
-  ) {
-    createInlineComment(
-      input: {
-        submission_id: $submission_id
-        content: $content
-        style_criteria: $style_criteria
-        from: $from
-        to: $to
-        reply_to_id: $reply_to_id
-        parent_id: $parent_id
-      }
-    ) {
-      id
-      inline_comments(trashed: WITH) {
-        style_criteria {
-          name
-          icon
-        }
-        ...commentFields
-        replies(trashed: WITH) {
-          ...commentFields
-          parent_id
-          reply_to_id
-        }
-      }
-    }
-  }
-`)
-
-graphql(`
-  mutation UpdateOverallComment(
-    $submission_id: ID!
-    $comment_id: ID!
-    $content: String!
-  ) {
-    updateOverallComment(
-      input: {
-        submission_id: $submission_id
-        comment_id: $comment_id
-        content: $content
-      }
-    ) {
-      id
-      created_by {
-        ...relatedUserFields
-      }
-      overall_comments(trashed: WITH) {
-        ...commentFields
-        replies(trashed: WITH) {
-          reply_to_id
-          parent_id
-          ...commentFields
-        }
-      }
-    }
-  }
-`)
-
-graphql(`
-  mutation UpdateInlineComment(
-    $submission_id: ID!
-    $comment_id: ID!
-    $content: String!
-    $style_criteria: [ID!]
-  ) {
-    updateInlineComment(
-      input: {
-        submission_id: $submission_id
-        comment_id: $comment_id
-        content: $content
-        style_criteria: $style_criteria
-      }
-    ) {
-      id
-      created_by {
-        ...relatedUserFields
-      }
-      inline_comments(trashed: WITH) {
-        ...commentFields
-        style_criteria {
-          name
-          icon
-        }
-        replies(trashed: WITH) {
-          reply_to_id
-          parent_id
-          ...commentFields
-        }
-      }
-    }
-  }
-`)
-</script>
-
 <script setup lang="ts">
 import { ref, computed, type ComputedRef } from "vue"
 import { useEditor, EditorContent } from "@tiptap/vue-3"
